+++ src/main/resources/application-prod.properties
@@ -9,6 +9,7 @@ spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/shiyizhonghua?serverTimezone=GMT%2B8&characterEncoding=utf-8
 spring.datasource.username=root
 spring.datasource.password=ZHUaowei++0521__
+
 # mongodb
 spring.data.mongodb.uri=mongodb://remote:123456@localhost:27017/shiyizhonghua
