+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ spring.web.resources.static-locations=classpath:/META-INF/resources/,classpath:/
 spring.thymeleaf.prefix=classpath:/templates/
 
 # mvc配置：mvc接收参数时对日期进行格式化
-spring.mvc.date-format=yyyy-MM-dd HH:mm:ss
+spring.mvc.format.date=yyyy-MM-dd HH:mm:ss
 
 # json格式化配置：jackson对响应回去的日期参数进行格式化
 spring.jackson.date-format=yyyy-MM-dd HH:mm:ss
