+++ src/main/resources/static/css/user/login.css
@@ -1,0 +1,155 @@
+*{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+body{
+    display: flex;
+    margin: 10% auto;
+    padding: 10% auto;
+    font-family: montserrat;
+    background-color:rgb(240,239,226);
+    width: 40rem;
+    height: 25rem;
+    overflow: hidden;
+}
+.center{
+    position: relative;
+    width: 70%;
+    height: 100%;
+    background:rgb(240,239,226);
+    border-radius: 10px 0 0 10px;
+    float:left;
+    border-top: 1px solid silver;
+    border-left: 1px solid silver;
+    border-bottom: 1px solid silver;
+}
+.center h1{
+    padding: 0 0 0px 30px;
+    border-bottom: 0px solid silver;
+}
+.center form{
+    padding: 0 40px;
+    box-sizing: border-box;
+}
+form .txt_field{
+    position: relative;
+    border-bottom: 2px solid #adadad;
+    margin: 30px 0;
+}
+.txt_field input{
+    width: 100%;
+    padding: 0 5px;
+    height: 40px;
+    font-size: 16px;
+    border: none;
+    background: none;
+    outline: none;
+}
+.txt_field label{
+    position:absolute;
+    top: 50%;
+    left: 5px;
+    color: #adadad;
+    transform: translateY(-50%);
+    font-size: 16px;
+    pointer-events: none;
+    transition: .5s;
+}
+.txt_field span::before{
+    content:'';
+    position:absolute;
+    top: 40px;
+    left: 0;
+    width:0%;
+    height:2px;
+    background: #2691d9;
+    transition: .5s;
+}
+.txt_field input:focus ~label,
+.txt_field input:valid ~label{
+    top: -5px;
+    color: #2691d9;
+}
+.txt_field input:focus ~span::before,
+.txt_field input:valid ~span::before{
+    width:100%;
+}
+.pass{
+    margin: -5px 0 20px 5px;
+    color:#a6a6a6;
+    cursor: pointer;
+    text-align: right;
+}
+.pass:hover{
+    text-decoration: underline;
+}
+input[type="submit"]{
+    width: 100%;
+    height: 50px;
+    border: 1px solid;
+    background:rgb(240,239,226);
+    border-radius: 15px;
+    font-size: 18px;
+    color: #000;
+    font-weight: 700;
+    cursor: pointer;
+    outline: none;
+}
+input[type="submit"]:hover{
+    border-color: #a6a6a6;
+    transition: .5s;
+}
+.zhuce_link{
+    margin:30px 0;
+    float: left;
+    font-size: 16px;
+    color: #a6a6a6;
+}
+.zhuce_link a{
+    color: #a6a6a6;
+    text-decoration: none;
+}
+.zhuce_link a:hover{
+    text-decoration: underline;
+}
+.yanzheng_link{
+    margin:30px 0;
+    float: right;
+    font-size: 16px;
+    color: #a6a6a6;
+}
+.yanzheng_link a{
+    color: #a6a6a6;
+    text-decoration: none;
+}
+.yanzheng_link a:hover{
+    text-decoration: underline;
+}
+.right{
+    width: 30%;
+    background:rgb(240,239,226);
+    border-radius: 0 10px 10px 0;
+    float:left;
+    border-top: 1px solid silver;
+    border-right: 1px solid silver;
+    border-bottom: 1px solid silver;
+}
+.right form{
+    padding: 5rem 20px;
+    box-sizing: border-box;
+}
+.qr{
+    margin: 10px 20px 20px 5px;
+    color:#a6a6a6;
+    cursor: pointer;    
+    text-align: right;
+}
+.qr:hover{
+    text-decoration: underline;
+}
+
+
+
+
+
